uint32v2

+++ src/scripts/export_yolo11.py
@@ -19,7 +19,9 @@ class DeepStreamOutput(nn.Module):
         x = x.transpose(1, 2)
         boxes = x[:, :, :4]
         scores, classes = torch.max(x[:, :, 4:], 2, keepdim=True)
-        classes = classes.float()
+        classes = classes.to(torch.int32)
+
+
         return boxes, scores, classes
